Stronger typing.

+++ mpeg2dec/doc/sample1.c
@@ -75,7 +75,7 @@ static void sample1 (FILE * mpgfile)
     }
     info = mpeg2_info (decoder);
 
-    size = -1;
+    size = (size_t)-1;
     do {
 	state = mpeg2_parse (decoder);
 	switch (state) {
