+++ mpeg2dec/doc/sample3.c
@@ -78,7 +78,7 @@ static void sample3 (FILE * mpgfile)
     }
     info = mpeg2_info (decoder);
 
-    size = -1;
+    size = (size_t)-1;
     do {
 	state = mpeg2_parse (decoder);
 	switch (state) {
